+++ app/src/commonMain/kotlin/com/hulkdx/findprofessional/app/di/AppModule.kt
@@ -1,6 +1,7 @@
 package com.hulkdx.findprofessional.app.di
 
 import com.hulkdx.findprofessional.app.di.module.apiModule
+import com.hulkdx.findprofessional.app.di.module.debugModule
 import com.hulkdx.findprofessional.core.storage.datastoreModule
 import com.hulkdx.findprofessional.feature.authentication.login.loginModule
 import com.hulkdx.findprofessional.feature.authentication.signup.signUpModule
@@ -9,6 +10,7 @@ import com.hulkdx.findprofessional.libs.navigation.decompose.decomposeModule
 
 val appModule
     get() = listOf(
+        debugModule(),
         apiModule,
         datastoreModule,
         decomposeModule,
