+++ app/src/commonTest/kotlin/com/hulkdx/findprofessional/app/test/screen/login/LoginScreenTest.kt
@@ -3,8 +3,8 @@
 package com.hulkdx.findprofessional.app.test.screen.login
 
 import androidx.compose.ui.test.ExperimentalTestApi
-import com.hulkdx.findprofessional.app.config.api.InMemoryApi
 import com.hulkdx.findprofessional.app.test.runAppUiTest
+import com.hulkdx.findprofessional.app.test.utils.setUser
 import kotlin.test.Test
 
 class LoginScreenTest {
@@ -20,7 +20,7 @@ class LoginScreenTest {
 
     @Test
     fun performLogin() = runAppUiTest {
-        InMemoryApi.setUser("test@email.com", "somepass")
+        setUser("test@email.com", "somepass")
 
         launchLoginScreen(this) {
             typeEmail("test@email.com")
