+++ app/src/commonTest/kotlin/com/hulkdx/findprofessional/app/test/utils/KoinUtils.kt
@@ -1,0 +1,7 @@
+package com.hulkdx.findprofessional.app.test.utils
+
+import org.koin.mp.KoinPlatformTools
+
+inline fun <reified T : Any> get(): T {
+    return KoinPlatformTools.defaultContext().get().get<T>()
+}
